// src/shared/types/validation.types.ts
export enum EValidationTypes {
  REQUIRED = 'REQUIRED',
  ONLY_LETTERS = 'ONLY_LETTERS',
  ONLY_NUMBERS = 'ONLY_NUMBERS',
  EMAIL_FORMAT = 'EMAIL_FORMAT',
  MIN_CHARACTERS = 'MIN_CHARACTERS',
  MAX_CHARACTERS = 'MAX_CHARACTERS',
  PHONE_NUMBER_FORMAT = 'PHONE_NUMBER_FORMAT',
  PASSWORD_FORMAT = 'PASSWORD_FORMAT',
  NO_SPECIAL_CHARACTERS = 'NO_SPECIAL_CHARACTERS',
  MATCH_REGEXP = 'MATCH_REGEXP',
  ENUM = 'ENUM',
  UNKNOWN_ERROR = 'UNKNOWN_ERROR'
}

export enum EApiResponse {
  NOT_AUTHORIZED = 'NotAuthorizedException'
}